requirements.txt
feedparser
python-dateutil
flask
gunicorn
psycopg2
